Web: Files: created withFileaActions HOC, refactoring SimpleFilesRow
