+++ products/ASC.Files/Client/src/components/pages/Home/Section/Body/sub-components/icons.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const EncryptedFileIcon = styled.div`
+  background: url("images/security.svg") no-repeat 0 0 / 16px 16px transparent;
+  height: 16px;
+  position: absolute;
+  width: 16px;
+  margin-top: 14px;
+  margin-left: ${(props) => (props.isEdit ? "40px" : "12px")};
+`;
